🚚 refactor: Change package name
