+++ application/src/main/kotlin/io/github/comstering/config/Swagger2Config.kt
@@ -1,4 +1,4 @@
-package com.example.demo.config
+package io.github.comstering.config
 
 import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.info.Info
